guice/src/main/java/com/impaq/ic/guice/named/player/MusicPlayerBean.java
package com.impaq.ic.guice.named.player;

import javax.inject.Inject;
import javax.inject.Named;

import com.impaq.ic.guice.named.listgenerator.PlayListGeneratorService;

public class MusicPlayerBean implements MusicPlayerService {

	private final PlayListGeneratorService playListGeneratorService;

	@Inject
	public MusicPlayerBean(
			@Named("Rock") PlayListGeneratorService playListGeneratorService) {
		if (playListGeneratorService == null) {
			throw new IllegalArgumentException(
					"PlayListGeneratorService must not be null");
		}
		this.playListGeneratorService = playListGeneratorService;
	}

	@Override
	public String getCurrentPlayedSong() {
		return playListGeneratorService.createPlayList();
	}

}
